move event fetching
