+++ core/src/vaults.ts
@@ -422,6 +422,15 @@ export const openVaultWithProxiedContractAndDrawDebt = async (
     const transactionData = method + encodedArgs.substring(2);
     const target = (await getContract(network, `PROXY_ACTIONS_${proxyType}`)).address;
     await proxyContract['execute(address,bytes)'](target, transactionData);
+
+    const registry = await getContract(network, 'CDP_REGISTRY');
+    const filter = registry.filters.NewCdpRegistered(null, proxyAddress, null);
+    const vaultOpenEvents = await registry.queryFilter(filter);
+    const vaultId = vaultOpenEvents[vaultOpenEvents.length - 1].args?.cdp;
+    if (!vaultId) {
+        throw new Error(`Failed to find event with opened vault id for ${collateralType}`);
+    }
+    return vaultId;
 };
 
 export const openVault = async (network: string, ownerAddress: string, collateralType: CollateralType) => {
